Fix CacheableMetadataStore test race conditions
The CacheableMetadataStore uses a seprate thread to listen for changes to the metadata store, and updates accordingly. This can cause a race condition in the tests if the cache has not yet updated.

This fixes a few missing checks where we didn't guarantee that the cache had updated prior to doing asserts.