Register Indexer in ChunkManager (#110)
Register Indexer during ChunkManager initialization.

ChunkManager initialization is not the ideal place to register the indexer, but doing so in this diff since it involves quite a few changes in the code. Once this diff is committed, will move the indexer registration closer to the snapshot registration. This diff also helps us make parallel progress on distributed search.

I have also refactored the code so most unit tests don't have a dependency on KaldbConfigUtil.initEmptyIndexerConfig anymore. This results in cleaner seperation of code and initialization of tests via ChunkManagerUtil.