Updated CachedMetadataImpl to work for single node. (#134)
Renamed CachedMetadataImpl to ZooKeeperCachedMetadataImpl to clarify that this is a ZK class and not a class in Kaldb metadata.

Rewrote the CachedMetadataImpl class to also work with caching a node and its children. This was a bigger change than expected since we need to a different curator API to get access both parent and child node notifications.

I also fixed a couple of exceptions and a bug in this diff.

Currently, when we create the cache, we already expect the node to be present. If not, the node is created with an empty data. However, with the new notification API, this causes a protobuf parse exception, since the new node is created with an empty string. So, I updated all the unit tests to create a node first. If needed, we can update the API with the initial value of the node, but none of our use cases need that right now.

Currently, we pass in executor service to the curator cache builder but that is a deprecated API that results in a warning. So, I removed the code that passes in executor service to the cache which simplifies the code. This makes our code more dependent on ZK 3.6.