Index registration (#138)
Register the indexer with snapshot store and search metadata store in Kaldb.

Until now, we don't pass in all the parameters needed to construct a proper snapshot into Kaldb. So, as part of this diff, we pass in those arguments also.

Added new unit tests where needed and updated existing tests.