Cleanup stale indexer nodes and recovery task creation.  (#176)
Currently, when an indexer restarts we leave stale metadata in ZK. This class cleans up stale indexer metadata. In addition, there is data loss for the data that is indexed but not uploaded to S3.

In this diff, we will implement stale live snapshot cleanup. In addition, we will determine the last offset in the durable metadata store and optionally create a recovery task if we can't catch up.