Recovery indexer. (#252)
This diff adds the logic for the recovery indexer.

A recovery indexer performs a recovery task and uploads a chunk to S3. To support this, we added a new chunk implementation called RecoveryChunkImpl that provides a ReadWriteChunk api, but doesn't publish live snapshots for the query service to query.

A RecoveryChunkManager implements a chunk manager for the recovery task.

I have also added a new parallel kafka consumer, that can ingest chunk data in parallel into the chunk manager as oppsed to the single threaded ingestion in the IndexingChunkManager.

Updated the RecoveryService to use the new parallel kafka consumer, recovery chunk manager and recovery chunk implementation to implement recovery functionality.