Add a `ChunkDownloader` interface that allows us to download all the chunk data. (#315)
Currently, the chunk downloading logic is hardcoded to an S3 API call. Add an abstraction around the snapshot downloading logic so we can plugin in different backends and also employ various downloading and rate limiting strategies.

Added a ChunkDownloader interface that provides this abstraction. In future, we can wrap this class with a rate limiter so we can easily rate limit the number of parallel rate limit calls.

Existing tests should validate the code, so no additional tests are needed.