Rename `ServiceMetadata` to `DatasetMetadata` (#341)
Similarly rename ServicePartitionMetadata to DatasetPartitionMetadata.

Updated all code references. Now we don't have different meanings for the word service in the code. Service is now called dataset.

We still need to change the ZK path for DatasetMetadata, but that's a small change we can do when we are re-provisioning the cluster from scratch.