Pass in a list of chunks instead of a chunk id in SearchQuery. (#359)
Pass in list of chunks instead of a single chunk id in the SearchQuery so we can search multiple chunks in a single call.

Also, update the search logic to only search the given chunks when chunk ids are specified. Added additional unit test for this code. Also, updated the unit test to query via the grpc query layer so we can validate end to end.