Add SchemaAwareLogDocumentBuilder to Kaldb. (#376)
In KalDB, the logs are ingested by a LogDocumentBuilder. However, the current document builder enforces types on known fields but any fields whose types are not pre-defined are ingested as Strings. This field conflict handling policy is not ideal for 2 reasons: (a) on dynamically ingested numeric fields, we can't do any range or numeric queries and (b) storing numbers as strings is also less efficient.

So, the SchemaAwareLogDocumentBuilder is a new LogDocumentBuilder that ingests the documents in a schema aware manner. It can take a fixed schema as before. In addition, for all dynamic fields, it also preserves the type of the field. If there is a field conflict, there is a field conflict policy that is applied to resolve the error. Currently, we have implemented 4 ways to handle field conflicts and also added unit tests to test the same. Added new tests for this functionality.

The overall work for integrating this schema document builder is being done in this PR: #369. I separated the code into a smaller diff for easier review.