Add index type to `CacheSlotMetadata`. (#500)
Added IndexType field to CacheSlotMetadata tasks so we can update the assignment logic in future to only assign snapshots of specific type to slots that support that type. I will change the assignment logic in a future diff. For now, only metadata is updated.

While making this change I also noticed that we update this CacheSlotState from multiple parts of our code and it's easy to miss copying over some fields that could cause subtle bugs. So, this diff also refactors that code to consolidate it to CacheSlotMetadataStore and added unit tests.

I think the state machine transition logic can be better designed in the future. So, left a TODO in CacheSlotMetadata to minimize the changes in a single diff.